refactor: `AttendanceHistoryItem` 모델 구조 통합 및 관련 데이터 전달 방식 개선
- `GameTeam` 및 `GameScoreBoard` 모델을 `AttendanceHistoryItem`의 내부 클래스로 이동 및 통합
- `AttendanceHistoryItem` 및 내부 클래스들에 `Serializable` 어노테이션 추가
- `AttendanceHistoryScreen`과 `AttendanceDetailScreen` 간의 데이터 전달 방식을 개별 필드(`id`, `date`)에서 `AttendanceHistoryItem` 객체 전체 전달 방식으로 변경
- `Route.AttendanceHistoryDetail` 네비게이션 인자를 `AttendanceHistoryItem`으로 수정
- `CheckInUiMapper`에서 변경된 `AttendanceHistoryItem` 내부 클래스 구조에 맞게 매핑 로직 수정
- 불필요해진 개별 모델 파일(`GameTeam.kt`, `GameScoreBoard.kt`) 삭제
