+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/navigation/NavigationRoot.kt
@@ -17,6 +17,7 @@ import androidx.navigation3.runtime.NavKey
 import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.ui.NavDisplay
 import com.yagubogu.ui.attendance.detail.AttendanceDetailScreen
+import com.yagubogu.ui.attendance.model.AttendanceHistoryItem
 import com.yagubogu.ui.badge.component.BadgeScreen
 import com.yagubogu.ui.livetalk.chat.LivetalkChatScreen
 import com.yagubogu.ui.login.LoginScreen
@@ -37,7 +38,6 @@ import com.yagubogu.ui.util.slidePopTransition
 import com.yagubogu.ui.util.slidePredictivePopTransition
 import com.yagubogu.ui.util.slidePushTransition
 import com.yagubogu.ui.util.snackbarPadding
-import kotlinx.datetime.LocalDate
 import org.koin.compose.viewmodel.koinViewModel
 import org.koin.core.parameter.parametersOf
 
@@ -88,8 +88,8 @@ fun NavigationRoot(
                         onLivetalkItemClick = { gameId: Long, isVerified: Boolean ->
                             rootNavigator.navigate(Route.LivetalkChat(gameId, isVerified))
                         },
-                        onAttendanceHistoryItemClick = { id: Long, date: LocalDate ->
-                            rootNavigator.navigate(Route.AttendanceHistoryDetail(id, date))
+                        onAttendanceHistoryItemClick = { item: AttendanceHistoryItem ->
+                            rootNavigator.navigate(Route.AttendanceHistoryDetail(item))
                         },
                     )
                 }
@@ -181,8 +181,7 @@ fun NavigationRoot(
                 }
                 entry<Route.AttendanceHistoryDetail> { key: Route.AttendanceHistoryDetail ->
                     AttendanceDetailScreen(
-                        checkInId = key.id,
-                        date = key.date,
+                        item = key.attendanceItem,
                         onBackClick = { rootNavigator.goBack() },
                     )
                 }
