+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/navigation/model/Route.kt
@@ -1,6 +1,7 @@
 package com.yagubogu.ui.navigation.model
 
 import androidx.navigation3.runtime.NavKey
+import com.yagubogu.ui.attendance.model.AttendanceHistoryItem
 import com.yagubogu.ui.ranking.model.RankingType
 import kotlinx.datetime.LocalDate
 import kotlinx.serialization.Serializable
@@ -42,7 +43,6 @@ sealed interface Route : NavKey {
 
     @Serializable
     data class AttendanceHistoryDetail(
-        val id: Long,
-        val date: LocalDate,
+        val attendanceItem: AttendanceHistoryItem,
     ) : Route
 }
